module compiling

+++ edgelet/edgelet-core/src/module.rs
@@ -54,7 +54,7 @@ pub struct ModuleRuntimeState {
     started_at: Option<DateTime<Utc>>,
     finished_at: Option<DateTime<Utc>>,
     image_id: Option<String>,
-    pid: Option<i64>,
+    pid: Option<i32>,
 }
 
 impl Default for ModuleRuntimeState {
@@ -126,11 +126,11 @@ impl ModuleRuntimeState {
         self
     }
 
-    pub fn pid(&self) -> Option<i64> {
+    pub fn pid(&self) -> Option<i32> {
         self.pid
     }
 
-    pub fn with_pid(mut self, pid: Option<i64>) -> Self {
+    pub fn with_pid(mut self, pid: Option<i32>) -> Self {
         self.pid = pid;
         self
     }
